交易记录导入优化

+++ src/main/java/cn/aixuxi/ledger/entity/BaseEntity.java
@@ -10,7 +10,7 @@ public class BaseEntity implements Serializable {
     /**
      * 创建人
      */
-    private Integer createUser;
+    private Long createUser;
     /**
      * 创建时间
      */
@@ -18,7 +18,7 @@ public class BaseEntity implements Serializable {
     /**
      * 更新人
      */
-    private Integer updateUser;
+    private Long updateUser;
     /**
      * 更新时间
      */
